\name{JSSim_Meta}
\docType{data}
\alias{JSSim_Meta}
\title{Meta File for Simulated Datasets}
\description{
This data set contains the links and description for the normal and tumor simulated data sets
}
\usage{JSSim_Meta}
\format{A matrix with 3 columns and 4 rows, and a header line}
\keyword{datasets}