\name{JSSim_NormalSim2}
\docType{data}
\alias{JSSim_NormalSim2}
\title{Simulated normal sample dataset 2}
\description{
This data set contains a second set of simulated reads of a truncated chromosome from a normal sample
}
\usage{JSSim_NormalSim2}
\format{A matrix with 3 columns and 15206 rows}
\keyword{datasets}