\name{JSSim_SpikeMat}
\docType{data}
\alias{JSSim_SpikeMat}
\title{True Signal Spike for the Simulated Dataset}
\description{
This data set gives the true signal spike location and strength for the simulated datasets
}
\usage{JSSim_SpikeMat}
\format{A matrix with 5 columns and rows}
\keyword{datasets}