\name{JSSim_TumorSim1}
\docType{data}
\alias{JSSim_TumorSim1}
\title{Simulated Tumor sample dataset 1}
\description{
This data set contains simulated reads of a truncated chromosome from a Tumor sample, with spiked in signals
}
\usage{JSSim_TumorSim1}
\format{A matrix with 3 columns and 16452 rows}
\keyword{datasets}