\name{JSSim_TumorSim2}
\docType{data}
\alias{JSSim_TumorSim2}
\title{Simulated Tumor sample dataset 2}
\description{
This data set contains a second set of simulated reads of a truncated chromosome from a Tumor sample, with spiked in signals
}
\usage{JSSim_TumorSim2}
\format{A matrix with 3 columns and 16225 rows}
\keyword{datasets}